+++ server/index.js
@@ -2,19 +2,15 @@ import express from "express";
 import cors from "cors";
 import morgan from "morgan";
 import authRouter from "./router/auth.js";
-import festaRouter from './router/c_festa.js';
-import placeRouter from './router/c_place.js';
+import cultureRouter from './router/culture.js';
 import { config } from "./config.js";
-import { sequelize } from "./db/database.js"
-// import mainController from "./controller/main.js"
+import { sequelize } from "./db/database.js";
 import mainRouter from "./router/main.js"
 import { EventEmitter } from 'events';
-import { initSocket } from './connection/socket.js'
-
 const bus = new EventEmitter();
 bus.setMaxListeners(20);
-
 const app = express();
+
 const corsOption = {
     origin: config.cors.allowedOrigin,
     optionsSuccessStatus: 200
@@ -27,10 +23,7 @@ app.use(morgan("tiny"));
 app.use("/auth", authRouter);
 app.use('/mypage', authRouter)
 app.use('/', mainRouter);
-app.use('/festa', festaRouter);
-app.use('/place', placeRouter);
-
-
+app.use('/culture', cultureRouter);
 
 app.use((req, res, next) => {
     res.sendStatus(404);
@@ -42,7 +35,5 @@ app.use((error, req, res, next) => {
 });
 
 sequelize.sync().then(() => {
-    console.log(`서버가 시작되었습니다: ${new Date()}`)
     const server = app.listen(config.host.port);
-    initSocket(server);
 });
